fix(AlertService): separates simple & autoDismiss stories
